requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
apex==0.1
async-timeout==3.0.1
attrs==20.3.0
blessed==1.18.1
blis==0.7.7
cachetools==4.2.1
catalogue==2.0.7
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
contextvars==2.4
crcmod==1.7
cryptography==3.4.8
cycler==0.11.0
cymem==2.0.6
dataclasses==0.8
datasets==2.4.0
decorator==5.1.1
deepspeed==0.3.16
dill==0.3.4
docopt==0.6.2
fairscale==0.3.3
faiss==1.7.2
faiss-cpu @ file:///data/guanjian/rept/wp/eval/faiss_cpu-1.7.2-cp36-cp36m-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
filelock==3.0.12
fire==0.4.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.7
gitdb2==4.0.2
GitPython==3.1.14
google-auth==1.28.1
google-auth-oauthlib==0.4.4
graphviz==0.19.1
grpcio==1.37.0
huggingface-hub==0.4.0
idna==2.10
idna-ssl==1.1.0
immutables==0.18
importlib-metadata==3.10.0
importlib-resources==5.4.0
jieba==0.42.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
jsonlines==3.0.0
kara-storage==2.1.1
kiwisolver==1.3.1
langcodes==3.3.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mauve-text==0.3.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
multiprocess==0.70.12.2
murmurhash==1.0.7
ninja==1.10.0.post2
nltk==3.6.1
numpy==1.19.5
oauthlib==3.1.0
objgraph==3.5.0
orjson==3.6.1
oss2==2.15.0
packaging==20.9
paddlepaddle-tiny==1.6.1
panda==0.3.1
pandas==1.1.5
pathy==0.6.1
patsy @ file:///data/guanjian/rept/wikitext-103/bys/corr/patsy-0.5.3-py2.py3-none-any.whl
Pillow==8.2.0
pipreqs==0.4.11
portalocker==2.0.0
preshed==3.0.6
protobuf==3.15.8
psutil==5.8.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-ignite==0.4.8
pytorch-lightning==1.2.7
pytz==2021.3
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
responses==0.17.0
rouge-score==0.0.4
rsa==4.7.2
sacrebleu==1.5.1
sacremoses==0.0.44
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sentencepiece==0.1.95
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smart-open==5.2.1
smmap==4.0.0
some-package==0.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
statsmodels @ file:///data/guanjian/rept/wikitext-103/bys/corr/statsmodels-0.12.2-cp36-cp36m-manylinux1_x86_64.whl
tensorboard==1.15.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.10.1
torch==1.8.1
torchmetrics==0.2.0
torchvision==0.9.1
tqdm==4.64.1
transformers==4.6.1
typer==0.4.1
typing-extensions==3.7.4.3
ujson==4.3.0
urllib3==1.26.4
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==1.2.0
xxhash==3.0.0
yarg==0.1.9
yarl==1.6.3
zipp==3.4.1
